Fix(script.js): refactor and document the code
Breaking Change:

The connection is established with the openWeatherMap API and the necessary information is extracted and then included in the application interface.

Note this: a promise is used for the fetch

A second button was included to return to the initial state of the sale, this acts as a 'close' for the window once the search is done.

The logic and general design of the application was improved and some bugs that it had when interacting with the application were fixed.